backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
pillow==10.1.0
openai-whisper==20231117
ffmpeg-python==0.2.0
python-magic-bin==0.4.14
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sentencepiece==0.1.99
protobuf==3.20.3
accelerate==0.25.0

# Object Detection
ultralytics==8.0.227

# Report Generation
reportlab==4.0.7
markdown==3.5.1

# WebSocket Support
websockets==12.0

# Job Queue (Optional - for batch processing)
celery==5.3.4
redis==5.0.1

# Visualization
wordcloud==1.9.3
matplotlib==3.8.2

# Authentication
python-multipart==0.0.6
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring & Logging
python-json-logger==2.0.7
